+++ src/components/MainSidebar/YourDimensionsPanel/YourDimensionsPanel.js
@@ -1,7 +1,9 @@
+import { useCachedDataQuery } from '@dhis2/analytics'
 import i18n from '@dhis2/d2-i18n'
 import { Input } from '@dhis2/ui'
 import PropTypes from 'prop-types'
 import React, { useState } from 'react'
+import { DERIVED_USER_SETTINGS_DISPLAY_NAME_PROPERTY } from '../../../modules/userSettings.js'
 import { useDebounce } from '../../../modules/utils.js'
 import { DimensionsList } from '../DimensionsList/index.js'
 import { useYourDimensions } from './useYourDimensions.js'
@@ -10,10 +12,12 @@ import styles from './YourDimensionsPanel.module.css'
 const YourDimensionsPanel = ({ visible }) => {
     const [searchTerm, setSearchTerm] = useState('')
     const debouncedSearchTerm = useDebounce(searchTerm)
+    const { userSettings } = useCachedDataQuery()
     const { loading, fetching, error, dimensions, setIsListEndVisible } =
         useYourDimensions({
             visible,
             searchTerm: debouncedSearchTerm,
+            nameProp: userSettings[DERIVED_USER_SETTINGS_DISPLAY_NAME_PROPERTY],
         })
 
     if (!visible) {