+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## [1.1.53](https://github.com/dhis2/line-listing-app/compare/v1.1.52...v1.1.53) (2022-04-26)
+
+
+### Bug Fixes
+
+* missing change from prev commit ([4b291ca](https://github.com/dhis2/line-listing-app/commit/4b291ca31362ec5fc6f5ebe3ef77ee15270b65b9))
+* remove shortname from options search as options only has displayName ([c23e882](https://github.com/dhis2/line-listing-app/commit/c23e8828a9c1ff4999c77075c295a54624dd4de1))
+* support shortName throughout the app (TECH-1118) [#108](https://github.com/dhis2/line-listing-app/issues/108) ([c55a1a1](https://github.com/dhis2/line-listing-app/commit/c55a1a1e20454f46b116c42fdb5d50b76a8bf62b))
+* use name instead of displayName in search ([745eeb4](https://github.com/dhis2/line-listing-app/commit/745eeb42b9d0de97d79db8531a6c888a726ad37e))
+* use shortname for program and dynamic/your dimensions ([f4c9053](https://github.com/dhis2/line-listing-app/commit/f4c9053185ce35a22b26341eb05e8c7465a37f8b))
+* use shortname for program when fetching AO ([b1e3016](https://github.com/dhis2/line-listing-app/commit/b1e3016f8a745be19ff3fe6b92ebcbc24e46c581))
+* use shortname for programs ([0599917](https://github.com/dhis2/line-listing-app/commit/05999176cb1100612cec03f89d381019c2f5ba9a))
+
 ## [1.1.52](https://github.com/dhis2/line-listing-app/compare/v1.1.51...v1.1.52) (2022-04-25)